requirements.txt
# Core dependencies
PyQt5>=5.15.0
requests>=2.31.0

# Network tools
python-nmap>=0.7.1
netifaces>=0.11.0
pyroute2>=0.7.3
scapy>=2.5.0

# System tools
psutil>=5.9.0

# Security tools
paramiko>=3.3.1  # For SSH functionality
pyOpenSSL>=23.2.0
cryptography>=41.0.0

# Wireless tools
wifi>=0.8.0
wireless>=0.3.0

# API Integrations
openai>=1.0.0  # Optional - direct API calls used instead

# Development tools
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional system dependencies (install via package manager):
# - hostapd
# - dnsmasq
# - apache2
# - vsftpd
# - openssh-server
# - aircrack-ng
# - cowpatty
# - lldpd
# - ethtool
# - network-manager
# - systemd-resolved 